Rename monitor.py to migration_monitor.py
