+++ FetchMigration/index_configuration_tool/tests/test_migration_monitor.py
@@ -6,7 +6,7 @@ import requests
 import responses
 from prometheus_client.parser import text_string_to_metric_families
 
-import monitor
+import migration_monitor
 from endpoint_info import EndpointInfo
 
 # Constants
@@ -20,12 +20,12 @@ TEST_PROMETHEUS_METRIC_STRING = "# HELP " + TEST_METRIC_NAME + " Unit Test Metri
     + TEST_METRIC_NAME + "{serviceName=\"unittest\",} " + str(TEST_METRIC_VALUE)
 
 
-class TestMonitor(unittest.TestCase):
+class TestMigrationMonitor(unittest.TestCase):
     @patch('requests.post')
     def test_shutdown(self, mock_post: MagicMock):
         expected_shutdown_url = TEST_ENDPOINT + "/shutdown"
         test_endpoint = EndpointInfo(TEST_ENDPOINT, TEST_AUTH, TEST_FLAG)
-        monitor.shutdown_pipeline(test_endpoint)
+        migration_monitor.shutdown_pipeline(test_endpoint)
         mock_post.assert_called_once_with(expected_shutdown_url, auth=TEST_AUTH, verify=TEST_FLAG)
 
     @patch('requests.get')
@@ -38,7 +38,7 @@ class TestMonitor(unittest.TestCase):
         type(mock_response).content = mock_content
         mock_get.return_value = mock_response
         # Test fetch
-        raw_metrics_list = monitor.fetch_prometheus_metrics(EndpointInfo(TEST_ENDPOINT))
+        raw_metrics_list = migration_monitor.fetch_prometheus_metrics(EndpointInfo(TEST_ENDPOINT))
         mock_get.assert_called_once_with(expected_url, auth=None, verify=True)
         self.assertEqual(1, len(raw_metrics_list))
         test_metric = raw_metrics_list[0]
@@ -65,17 +65,17 @@ class TestMonitor(unittest.TestCase):
         expected_val = int(TEST_METRIC_VALUE)
         test_input = list(text_string_to_metric_families(TEST_PROMETHEUS_METRIC_STRING))
         # Should fetch by suffix
-        val = monitor.get_metric_value(test_input, "metric")
+        val = migration_monitor.get_metric_value(test_input, "metric")
         self.assertEqual(expected_val, val)
         # No matching metric returns None
-        val = monitor.get_metric_value(test_input, "invalid")
+        val = migration_monitor.get_metric_value(test_input, "invalid")
         self.assertEqual(None, val)
 
-    @patch('monitor.shutdown_pipeline')
+    @patch('migration_monitor.shutdown_pipeline')
     @patch('time.sleep')
-    @patch('monitor.check_if_complete')
-    @patch('monitor.get_metric_value')
-    @patch('monitor.fetch_prometheus_metrics')
+    @patch('migration_monitor.check_if_complete')
+    @patch('migration_monitor.get_metric_value')
+    @patch('migration_monitor.fetch_prometheus_metrics')
     # Note that mock objects are passed bottom-up from the patch order above
     def test_run(self, mock_fetch: MagicMock, mock_get: MagicMock, mock_check: MagicMock, mock_sleep: MagicMock,
                  mock_shut: MagicMock):
@@ -88,7 +88,7 @@ class TestMonitor(unittest.TestCase):
         mock_check.side_effect = [False, True]
         # Run test method
         wait_time = 3
-        monitor.run(test_input, wait_time)
+        migration_monitor.run(test_input, wait_time)
         # Test that fetch was called with the expected EndpointInfo
         expected_endpoint_info = EndpointInfo(test_input.dp_endpoint, ('admin', 'admin'), False)
         self.assertEqual(2, mock_fetch.call_count)
@@ -126,17 +126,17 @@ class TestMonitor(unittest.TestCase):
 
     def test_check_if_complete(self):
         # If any of the optional values are missing, we are not complete
-        self.assertFalse(monitor.check_if_complete(None, 0, 1, 0, 2))
-        self.assertFalse(monitor.check_if_complete(2, None, 1, 0, 2))
-        self.assertFalse(monitor.check_if_complete(2, 0, None, 0, 2))
+        self.assertFalse(migration_monitor.check_if_complete(None, 0, 1, 0, 2))
+        self.assertFalse(migration_monitor.check_if_complete(2, None, 1, 0, 2))
+        self.assertFalse(migration_monitor.check_if_complete(2, 0, None, 0, 2))
         # Target count not reached
-        self.assertFalse(monitor.check_if_complete(1, None, None, 0, 2))
+        self.assertFalse(migration_monitor.check_if_complete(1, None, None, 0, 2))
         # Target count reached, but has records in flight
-        self.assertFalse(monitor.check_if_complete(2, 1, None, 0, 2))
+        self.assertFalse(migration_monitor.check_if_complete(2, 1, None, 0, 2))
         # Target count reached, no records in flight, but no prev no_part_count
-        self.assertFalse(monitor.check_if_complete(2, 0, 1, 0, 2))
+        self.assertFalse(migration_monitor.check_if_complete(2, 0, 1, 0, 2))
         # Terminal state
-        self.assertTrue(monitor.check_if_complete(2, 0, 2, 1, 2))
+        self.assertTrue(migration_monitor.check_if_complete(2, 0, 2, 1, 2))
 
 
 if __name__ == '__main__':
